Converted `$findMatchingRoute` to use regular expressions. For #18
The process starts in the Mapper, by converting the pattern into a regex
that gets passed along to the routes structure. When a request comes in,
the path gets compared against each route's regex. This is shorter code,
and probably more efficient, but unfortunately breaks all routes added
via wheels `addRoute` function.